requirements.txt
# calls platform.python_version()[:3] under the covers
python_version >= 3.6
numpy >= 1.17.2
pandas >= 0.25.3
shapely >= 1.6.4.post2
osmxtract >= 0.0.1
Pillow >= 6.2.1
matplotlib ~= 3.1.2
wheel ~= 0.33
requests ~= 2.22
